+++ src/main/java/ladder/controller/LadderController.java
@@ -4,12 +4,11 @@ import ladder.domain.Height;
 import ladder.domain.LadderGame;
 import ladder.domain.Ladders;
 import ladder.domain.Participants;
-import ladder.domain.RandomPointGenerationStrategy;
+import ladder.domain.strategy.RandomPointGenerationStrategy;
 import ladder.view.input.InputHeightView;
 import ladder.view.input.InputNamesView;
 import ladder.view.output.OutputLaddersView;
 import ladder.view.output.OutputNamesView;
-import ladder.view.output.OutputView;
 
 public class LadderController {
