+++ src/main/java/ladder/domain/LadderGame.java
@@ -3,6 +3,7 @@ package ladder.domain;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import ladder.domain.strategy.NextPointGenerationStrategy;
 
 
 public class LadderGame {
